pyproject.toml
[project]
name = "big-data-project"
version = "0.1.0"
description = "Mastodon streaming pipeline to Kafka"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.6.0",
    "python-dotenv>=1.0.0",
    "mastodon-py>=1.8.0",
    "flask>=3.1.2",
    "fastapi>=0.128.0",
    "pydoop>=2.0.0",
    "kafka-python>=2.3.0",
    "rich>=14.2.0",
    "apache-flink>=2.2.0",
    "elasticsearch>=8.0.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
